내 페이지 구현

+++ src/main/java/com/minidooray/gateway/account/domain/Account.java
@@ -9,7 +9,7 @@ import java.time.LocalDate;
 @Setter
 public class Account {
 
-    private Integer accountSeq;
+    private Long accountSeq;
 
     private String accountId;
